Split getstdlib in two files for better readability
